+++ src/config.ts
@@ -1,5 +1,5 @@
 import { i18n } from 'dateformat'
-import { dayNames, monthNames, timeNames } from './language/dateformat'
+import { dayNames, monthNames, timeNames } from '@language/dateformat'
 
 const config = {
     language: 'en-US'