+++ src/types/username.d.ts
@@ -1,0 +1,1 @@
+type AllowedUsernameValues = 'a-z' | 'A-Z' | '0-9' | '_' | '-' | '.'