+++ src/utils/validators/isUsernameValid.ts
@@ -1,5 +1,3 @@
-type AllowedValues = 'a-z' | 'A-Z' | '0-9' | '_' | '-' | '.'
-
 /**
  * Checks if username is valid
  *
@@ -16,7 +14,7 @@ type AllowedValues = 'a-z' | 'A-Z' | '0-9' | '_' | '-' | '.'
  * OU.isUsernameValid("ovni.dev", 3, 10, ['az','AZ','.']) // Return true
  * OU.isUsernameValid("dev", 3, 10) // Return true
  */
-const isUsernameValid = (username: string, minLength = 3, maxLength = 20, charactersAllowed: AllowedValues[] = []): boolean => {
+const isUsernameValid = (username: string, minLength = 3, maxLength = 20, charactersAllowed: AllowedUsernameValues[] = []): boolean => {
     let charactersRegexp = ''
     charactersAllowed.forEach((char: string) => { charactersRegexp += char })
     const validRegexp = new RegExp(`^${charactersRegexp ? '[' + charactersRegexp + ']' : '[a-z]'}{${minLength},${maxLength}}$`)